PSE3/requirements.txt
python-dotenv
googletrans==4.0.0-rc1
word2number
google-generativeai
